requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=2.1.0=pypi_0
accelerate=0.30.1=pypi_0
addict=2.4.0=pypi_0
aiohappyeyeballs=2.3.4=pypi_0
aiohttp=3.10.0=pypi_0
aiosignal=1.3.1=pypi_0
argcomplete=3.4.0=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
boto=2.49.0=pypi_0
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.8.30=hbcca054_0
cachetools=5.4.0=pypi_0
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.16.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
clinfo=3.0.21.02.21=h7f98852_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.2.1=pypi_0
crcmod=1.7=pypi_0
cryptography=43.0.0=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
docker-pycreds=0.4.0=pypi_0
easydict=1.13=pypi_0
einops=0.8.0=pypi_0
executing=2.0.1=pypi_0
facexlib=0.3.0=pypi_0
farama-notifications=0.0.4=pypi_0
fasteners=0.19=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
filterpy=1.4.5=pypi_0
fonttools=4.51.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
ftfy=6.2.0=pypi_0
future=1.0.0=pypi_0
gcloud=0.18.3=pypi_0
gcs-oauth2-boto-plugin=3.2=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-apitools=0.5.32=pypi_0
google-auth=2.17.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-reauth=0.1.1=pypi_0
googleapis-common-protos=1.63.2=pypi_0
gpustat=1.1.1=pypi_0
grpcio=1.63.0=pypi_0
gsutil=5.30=pypi_0
gymnasium=0.28.1=pypi_0
httplib2=0.20.4=pypi_0
huggingface-hub=0.23.0=pypi_0
icecream=2.1.3=pypi_0
idna=3.7=py39h06a4308_0
imageio=2.34.1=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jax-jumpy=1.0.0=pypi_0
jinja2=3.1.3=py39h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=14.1.0=h4852527_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
lightning-utilities=0.11.2=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.42.0=pypi_0
lmdb=1.4.1=pypi_0
lz4-c=1.9.4=h6a678d5_1
mako=1.3.5=pyhd8ed1ab_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.8.4=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
monotonic=1.6=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
multidict=6.0.5=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
numba=0.59.1=pypi_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
nvidia-ml-py=12.550.52=pypi_0
oauth2client=4.1.3=pypi_0
ocl-icd=2.3.2=h5eee18b_1
openai-clip=1.0.1=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pillow=10.3.0=py39h5eee18b_0
pip=24.0=py39h06a4308_0
platformdirs=4.2.1=pypi_0
pretty-errors=1.2.25=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pypi_0
pygments=2.18.0=pypi_0
pyiqa=0.1.11=pypi_0
pyopencl=2024.2.7=py39h85c637f_0
pyopengl=3.1.6=pyhd8ed1ab_1
pyopenssl=24.2.1=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.9=2_cp39
pytools=2024.1.14=pyhd8ed1ab_0
pytorch=2.0.0=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyu2f=0.1.5=pypi_0
pyyaml=6.0.1=pypi_0
rawpy=0.22.0=pypi_0
readline=8.2=h5eee18b_0
regex=2024.5.15=pypi_0
requests=2.31.0=py39h06a4308_1
retry-decorator=1.1.1=pypi_0
rich=13.7.1=pypi_0
rsa=4.7.2=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.22.0=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.1.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=py39h06a4308_0
shapely=2.0.4=pypi_0
siphash24=1.6=py39hcd6043d_1
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
thop=0.1.1-2209072238=pypi_0
tifffile=2024.5.10=pypi_0
timm=1.0.3=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.15.2=pypi_0
tomli=2.0.1=pypi_0
torchaudio=2.0.0=py39_cu118
torchmetrics=1.4.0=pypi_0
torchtriton=2.0.0=py39
torchvision=0.15.0=py39_cu118
tqdm=4.66.4=pypi_0
transformers=4.37.2=pypi_0
typing-extensions=4.11.0=py39h06a4308_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024.1=pypi_0
ultralytics=8.2.14=pypi_0
urllib3=2.1.0=py39h06a4308_1
wandb=0.17.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
yapf=0.40.2=pypi_0
yarl=1.9.4=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
